requirements.txt
Django==4.0
djangorestframework

python-dotenv
django-ckeditor-5
psycopg2-binary
factory_boy